% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_value_extract.R
\name{convert_to_numeric}
\alias{convert_to_numeric}
\title{Convert card value to numeric.}
\usage{
convert_to_numeric(card)
}
\arguments{
\item{card}{A character vector representing a single card.}
}
\value{
The numeric value of the card.
}
\description{
This function converts the value of a card to its numeric equivalent.
For example, '2' becomes 2, '3' becomes 3, 'T' becomes 10, 'J' becomes 11, etc.
}
\keyword{internal}
