% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker_leaderboard.R
\name{create_scoreboard}
\alias{create_scoreboard}
\title{Used to create a Scoreboard vector which consists of winner, pot amount, bet amount for each round}
\usage{
create_scoreboard(player_no, pot, bet)
}
\arguments{
\item{player_no}{A numeric vector having the player numbers}

\item{pot}{A numeric vector having the pot amounts}

\item{bet}{A numeric vector bet amounts which can't be more than pot amount}
}
\value{
scoreboard A custom vector usable to calculate leaderboard
}
\description{
Used to create a Scoreboard vector which consists of winner, pot amount, bet amount for each round
}
\examples{
create_scoreboard(c(1,2,3), c(5000,3000,4000), c(200,500,300))
}
