% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_hand.R
\name{determine_best_hand}
\alias{determine_best_hand}
\title{Determine the best hand.}
\usage{
determine_best_hand(hand)
}
\arguments{
\item{hand}{A character vector representing a poker hand.}
}
\value{
An integer indicating the rank of the hand.
}
\description{
This function determines the rank of a given poker hand, based on various combinations.
}
\keyword{internal}
