% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker_leaderboard.R
\name{format.scores}
\alias{format.scores}
\title{Custom Formatting of the scores class}
\usage{
\method{format}{scores}(x, ...)
}
\arguments{
\item{x}{A scores vector rcrd class}

\item{...}{additional parameters if required}
}
\value{
Neatly prints the scores board
}
\description{
Custom Formatting of the scores class
}
