% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flush_hand.R
\name{is_flush}
\alias{is_flush}
\title{Check if a hand is a flush.}
\usage{
is_flush(hand)
}
\arguments{
\item{hand}{A character vector representing a poker hand.}
}
\value{
A logical value indicating whether the hand is a flush.
}
\description{
This function checks if a hand is a flush, which is a hand that
contains five cards of the same suit.
}
\keyword{internal}
