% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_kind_hand.R
\name{is_four_of_a_kind}
\alias{is_four_of_a_kind}
\title{Check if a hand is a four of a kind.}
\usage{
is_four_of_a_kind(hand)
}
\arguments{
\item{hand}{A character vector representing a poker hand.}
}
\value{
A logical value indicating whether the hand is a four of a kind.
}
\description{
This function checks if a hand is a four of a kind, which is a hand that
contains four cards of the same rank.
}
\keyword{internal}
