% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_house_hand.R
\name{is_full_house}
\alias{is_full_house}
\title{Check if a hand is a full house.}
\usage{
is_full_house(hand)
}
\arguments{
\item{hand}{A character vector representing a poker hand.}
}
\value{
A logical value indicating whether the hand is a full house.
}
\description{
This function checks if a hand is a full house, which is a hand that
contains three cards of one rank and two cards of another rank.
}
\keyword{internal}
