% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_pair_hand.R
\name{is_pair}
\alias{is_pair}
\title{Check if a hand is a pair.}
\usage{
is_pair(hand)
}
\arguments{
\item{hand}{A character vector representing a poker hand.}
}
\value{
A logical value indicating whether the hand is a pair.
}
\description{
This function checks if a hand is a pair, which is a hand that
contains two cards of the same rank.
}
\keyword{internal}
