% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/royal_flush_hand.R
\name{is_royal_flush}
\alias{is_royal_flush}
\title{Check if a hand is a royal flush.}
\usage{
is_royal_flush(hand)
}
\arguments{
\item{hand}{A character vector representing a poker hand.}
}
\value{
A logical value indicating whether the hand is a royal flush.
}
\description{
This function checks if a hand is a royal flush, which is a hand that
contains five cards of the same suit with values '10', 'Jack', 'Queen', 'King',
and 'Ace'.
}
\details{
This function internally calls the is_flush function.
}
\keyword{internal}
