% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straight_hand.R
\name{is_straight}
\alias{is_straight}
\title{Check if a hand is a straight.}
\usage{
is_straight(hand)
}
\arguments{
\item{hand}{A character vector representing a poker hand.}
}
\value{
A logical value indicating whether the hand is a straight.
}
\description{
This function checks if a hand is a straight, which is a hand that
contains five cards of sequential rank.
}
\keyword{internal}
