% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_kind_hand.R
\name{is_three_of_a_kind}
\alias{is_three_of_a_kind}
\title{Check if a hand is a three of a kind.}
\usage{
is_three_of_a_kind(hand)
}
\arguments{
\item{hand}{A character vector representing a poker hand.}
}
\value{
A logical value indicating whether the hand is a three of a kind.
}
\description{
This function checks if a hand is a three of a kind, which is a hand that
contains three cards of the same rank.
}
\keyword{internal}
