% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_pair_hand.R
\name{is_two_pair}
\alias{is_two_pair}
\title{Check if a hand is a two pair.}
\usage{
is_two_pair(hand)
}
\arguments{
\item{hand}{A character vector representing a poker hand.}
}
\value{
A logical value indicating whether the hand is a two pair.
}
\description{
This function checks if a hand is a two pair, which is a hand that
contains two different pairs of cards of the same rank.
}
\keyword{internal}
