% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_hand.R
\name{is_valid_hand}
\alias{is_valid_hand}
\title{Checking valid hand input.}
\usage{
is_valid_hand(hand)
}
\arguments{
\item{hand}{A character vector representing a poker hand.}
}
\value{
A logical value indicating whether the hand is valid.
}
\description{
This function checks if a given input represents a valid poker hand.
}
\keyword{internal}
