% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker777-package.R
\docType{package}
\name{poker777-package}
\alias{poker777}
\alias{poker777-package}
\title{poker777: The Package Determines The Winner Of A Poker Game}
\description{
The package takes input from a user for cards in player 1's hand and player 2's hand and then determines which player is the winner by determining the players's rank according to the combination of cards and hence the lower rank wins the game.
}
\author{
\strong{Maintainer}: Harsh Jain \email{hjai0003@student.monash.edu}

}
\keyword{internal}
