% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chips_balance.R
\name{poker_balance_chips}
\alias{poker_balance_chips}
\title{Calculate the balance chips in hand after betting a certain rounds}
\usage{
poker_balance_chips(buy_in_limits, min_bet, player_buy_in, bet_n_win)
}
\arguments{
\item{buy_in_limits}{Minimum and Maximum chips to buy in a table}

\item{min_bet}{Minimum bet a player would need to place every round}

\item{player_buy_in}{Actual chips that player bought in}

\item{bet_n_win}{A list of betting amount and winning amount for each round. Input must be like list(c(bet1,win1),c(bet2,win2),...)}
}
\value{
A number indicating the outstanding chips in the player's hand at the last round.
}
\description{
Calculate the balance chips in hand after betting a certain rounds
}
\examples{

buy_in_limits <- c(5000, 20000)
min_bet <- 50
player_buy_in <- 10000
bet_n_win <- list(c(500, 5000), c(100, 0), c(200, 1000))
poker_balance_chips(buy_in_limits, min_bet, player_buy_in, bet_n_win)
}
