% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker_leaderboard.R
\name{poker_leaderboard}
\alias{poker_leaderboard}
\title{Leaderboard Generator. Based on the winners of each round and pot amount, arranges player positions and shows net amount won and rounds won by each player}
\usage{
poker_leaderboard(scoreboard)
}
\arguments{
\item{scoreboard}{A vector of type scores having the player no, pot amounts and bet amounts}
}
\description{
Leaderboard Generator. Based on the winners of each round and pot amount, arranges player positions and shows net amount won and rounds won by each player
}
\examples{
pot_wins <- create_scoreboard(c(1,2,3), c(5000,3000,4000), c(200,500,300))
poker_leaderboard(pot_wins)
}
