% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker_leaderboard.R
\name{vec_cast.data.frame.scores}
\alias{vec_cast.data.frame.scores}
\title{Custom method for casting the scores class to data frame class}
\usage{
vec_cast.data.frame.scores(x, to, ...)
}
\arguments{
\item{x}{A scores vector rcrd class}

\item{to}{A class that we need to convert to. By default a data frame}

\item{...}{Other parameters if required}
}
\value{
A data frame
}
\description{
Custom method for casting the scores class to data frame class
}
