% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker_leaderboard.R
\name{vec_cast.scores.data.frame}
\alias{vec_cast.scores.data.frame}
\title{Custom method for casting the data frame class to scores class}
\usage{
\method{vec_cast}{scores.data.frame}(x, to, ...)
}
\arguments{
\item{x}{A data frame to be converted to rcrd class}

\item{to}{A class that we need to convert to. By default scores}

\item{...}{additional parameters if required}
}
\value{
A rcrd vector
}
\description{
Custom method for casting the data frame class to scores class
}
