% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker_leaderboard.R
\name{vec_ptype2.data.frame.scores}
\alias{vec_ptype2.data.frame.scores}
\title{Custom method for prototyping the data frame class with scores class}
\usage{
\method{vec_ptype2}{data.frame.scores}(x, y, ...)
}
\arguments{
\item{x}{A data frame class}

\item{y}{A scores vector rcrd class}

\item{...}{additional parameters if required}
}
\value{
A data frame
}
\description{
Custom method for prototyping the data frame class with scores class
}
