% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker_leaderboard.R
\name{vec_ptype2.scores.data.frame}
\alias{vec_ptype2.scores.data.frame}
\title{Custom method for prototyping the scores class with data frame class}
\usage{
\method{vec_ptype2}{scores.data.frame}(x, y, ...)
}
\arguments{
\item{x}{A scores vector rcrd class}

\item{y}{A data frame class}

\item{...}{additional parameters if required}
}
\value{
A data frame
}
\description{
Custom method for prototyping the scores class with data frame class
}
